Fix code errors detected using Linaro GCC 6.1.1-rc1
Performed a build with the latest Linaro GCC toolchain explicitly to
take advantage of improved code error detection. These appeared as
warnings from the compiler.

For the OnePlus One (bacon).

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -57,9 +57,9 @@
 
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
-	int t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
-			(mask)) << (shift)), addr); \
+	uint64_t t = readl_relaxed(addr); \
+	writel_relaxed((t & ~(((uint64_t) mask) << (shift))) + (((v) & \
+			((uint64_t) mask)) << (shift)), addr); \
 } while (0)
